requirements.docker.txt
fastapi==0.136.1
uvicorn==0.46.0
sqlalchemy==2.0.49
qiskit==2.4.1
qiskit-ibm-runtime==0.46.1
python-dotenv==1.2.2
pydantic==2.13.3
